Extract IO-isolated core into separate crate
