+++ crates/config/Cargo.toml
@@ -2,14 +2,15 @@
 authors = ["lhvy <me@lhvy.dev>"]
 edition = "2018"
 license = "MIT OR Apache-2.0"
-name = "pipes-rs"
-version = "1.4.4"
+name = "config"
+version = "0.1.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 anyhow = "1.0.37"
-config = {path = "../config"}
-mimalloc = {version = "0.1.25", default-features = false}
-pipes-rs-core = {path = "../pipes-rs-core"}
-terminal = {path = "../terminal"}
+etcetera = "0.3.2"
+model = {path = "../model"}
+serde = "1.0.118"
+structopt = "0.3.21"
+toml = "0.5.8"
